requirements.txt
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
faker>=20.0.0